Extract RoomPrivate
